XWIKI-22323: Refactoring operation should wait for the Solr index to be empty before proceeding
* Move the ReadyIndicator to the store API.
