+++ xwiki-platform-core/xwiki-platform-search/xwiki-platform-search-solr/xwiki-platform-search-solr-api/src/main/java/org/xwiki/search/solr/internal/metadata/DefaultLinkStore.java
@@ -36,7 +36,6 @@ import org.apache.solr.common.SolrDocumentList;
 import org.xwiki.component.annotation.Component;
 import org.xwiki.link.LinkException;
 import org.xwiki.link.LinkStore;
-import org.xwiki.link.ReadyIndicator;
 import org.xwiki.model.EntityType;
 import org.xwiki.model.internal.reference.AbstractEntityReferenceResolver;
 import org.xwiki.model.reference.DocumentReference;
@@ -51,6 +50,7 @@ import org.xwiki.search.solr.internal.SolrClientInstance;
 import org.xwiki.search.solr.internal.SolrSearchCoreUtils;
 import org.xwiki.search.solr.internal.api.FieldUtils;
 import org.xwiki.search.solr.internal.api.SolrIndexer;
+import org.xwiki.store.ReadyIndicator;
 
 /**
  * @version $Id$
